+++ b2sdk/file_version.py
@@ -8,10 +8,8 @@
 #
 ######################################################################
 
-from typing import Optional
-
 from .encryption.setting import EncryptionSetting, EncryptionSettingFactory
-from .file_lock import FileRetentionSetting, LegalHold
+from .file_lock import FileRetentionSetting, LegalHold, NO_RETENTION_FILE_SETTING
 from .raw_api import SRC_LAST_MODIFIED_MILLIS
 
 
@@ -36,6 +34,7 @@ class FileVersion:
 
     __slots__ = [
         'id_',
+        'api',
         'file_name',
         'size',
         'content_type',
@@ -52,6 +51,7 @@ class FileVersion:
 
     def __init__(
         self,
+        api,
         id_,
         file_name,
         size,
@@ -60,14 +60,12 @@ class FileVersion:
         file_info,
         upload_timestamp,
         action,
-        content_md5=None,
-        server_side_encryption: Optional[EncryptionSetting] = None,  # TODO: make it mandatory in v2
-        file_retention: Optional[
-            FileRetentionSetting
-        ] = None,  # TODO: in v2 change the default value to NO_RETENTION_FILE_SETTING
-        legal_hold: Optional[LegalHold
-                            ] = None,  # TODO: in v2 change the default value to LegalHold.UNSET
+        content_md5,
+        server_side_encryption: EncryptionSetting,
+        file_retention: FileRetentionSetting = NO_RETENTION_FILE_SETTING,
+        legal_hold: LegalHold = LegalHold.UNSET,
     ):
+        self.api = api
         self.id_ = id_
         self.file_name = file_name
         self.size = size
@@ -93,6 +91,8 @@ class FileVersion:
             'fileName': self.file_name,
             'fileInfo': self.file_info,
             'legalHold': self.legal_hold.to_dict_repr() if self.legal_hold is not None else None,
+            'serverSideEncryption': self.server_side_encryption.as_dict(),
+            'fileRetention': self.file_retention.as_dict(),
         }
 
         if self.size is not None:
@@ -107,10 +107,6 @@ class FileVersion:
             result['contentSha1'] = self.content_sha1
         if self.content_md5 is not None:
             result['contentMd5'] = self.content_md5
-        if self.server_side_encryption is not None:  # this is for backward compatibility of interface only, b2sdk always sets it
-            result['serverSideEncryption'] = self.server_side_encryption.as_dict()
-        if self.file_retention is not None:  # this is for backward compatibility of interface only, b2sdk always sets it
-            result['fileRetention'] = self.file_retention.as_dict()
         return result
 
     def __eq__(self, other):
@@ -133,7 +129,7 @@ class FileVersionFactory(object):
     """
 
     @classmethod
-    def from_api_response(cls, file_version_dict, force_action=None):
+    def from_api_response(cls, api, file_version_dict, force_action=None):
         """
         Turn this:
 
@@ -188,6 +184,7 @@ class FileVersionFactory(object):
         legal_hold = LegalHold.from_file_version_dict(file_version_dict)
 
         return FileVersion(
+            api,
             id_,
             file_name,
             size,
@@ -203,8 +200,9 @@ class FileVersionFactory(object):
         )
 
     @classmethod
-    def from_response_headers(cls, headers):
+    def from_response_headers(cls, api, headers):
         return FileVersion(
+            api=api,
             id_=headers.get('x-bz-file-id'),
             file_name=headers.get('x-bz-file-name'),
             size=headers.get('content-length'),
@@ -213,6 +211,7 @@ class FileVersionFactory(object):
             file_info=None,
             upload_timestamp=headers.get('x-bz-upload-timestamp'),
             action=None,
+            content_md5=None,
             server_side_encryption=EncryptionSettingFactory.from_response_headers(headers),
             file_retention=FileRetentionSetting.from_response_headers(headers),
             legal_hold=LegalHold.from_response_headers(headers),